requirements-base.txt
# Tools
ipython==3.2.0

# General libs
pytz==2015.4
bleach==1.4.1
Markdown==2.6.2
Pillow==2.9.0
requests==2.7.0

# Google Drive API
google-api-python-client==1.4.1


# Django
Django==1.8.2
django-authtools==1.2.0
django-braces==1.8.0
django-extensions==1.5.5
django-floppyforms==1.4.1
